backend/tools/portionnet/requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
Pillow>=9.0.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Point cloud processing
open3d>=0.16.0

# Training utilities
tqdm>=4.62.0

# Data augmentation
albumentations>=1.3.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0

# Experiment tracking (optional)
tensorboard>=2.10.0
wandb>=0.13.0

# Excel file reading
openpyxl>=3.0.0
xlrd>=2.0.0
